Discrete without baselines working now.
Fixed a bug in 1.rward_to_go calc, and 2. probability normalization
Was using len(path) instead of len(path['reward']).
And in probablity calculation, was not subtracting the normalization
constant from the chosen logit. Since the build_mlp() function only
outputs the logits, the chosen probability logic needs to have the logic
for softmax.